Draft: esp-idf (ESP32) port

See merge request libssh/libssh-mirror!522

+++ src/bind_config.c
@@ -319,7 +319,7 @@ ssh_bind_config_parse_line(ssh_bind bind,
 
     /* Remove trailing spaces */
     for (len = strlen(s) - 1; len > 0; len--) {
-        if (! isspace(s[len])) {
+        if (! isspace((int)s[len])) {
             break;
         }
         s[len] = '\0';
