+++ src/config.c
@@ -766,7 +766,7 @@ ssh_config_parse_line(ssh_session session,
 
   /* Remove trailing spaces */
   for (len = strlen(s) - 1; len > 0; len--) {
-    if (! isspace(s[len])) {
+    if (! isspace((int)s[len])) {
       break;
     }
     s[len] = '\0';
